requirements.txt
alembic==1.13.2
fastapi==0.111.1
mysql-connector-python==9.0.0
mysqlclient==2.2.4
SQLAlchemy==2.0.31
uvicorn==0.30.3
pyjwt[crypto]==2.8.0
streaming-form-data==1.15.0
pytest==8.3.2
pytest-cov==5.0.0
starlette~=0.37.2
pydantic~=2.8.2
bcrypt~=4.2.0
typer==0.12.3
